Initial forecasts by multiple regression
Commented out sections in Rscripts need to be corrected, in particular the method for retrieving the forecasting data from the plots. This initial forecast illustrates strong oscillatory behaviour in the residuals, while the confidence intervals are probably too wide. Moving average and exponential smoothing techniques, and a different choice of predictors for regression are going to be tried next